Integrate changes from wikiteam
