+++ MANIFEST.in
@@ -1,2 +1,2 @@
 include README.md CHANGELOG
-graft wikitools
+graft wikitools3
