+++ README.md
@@ -1,4 +1,4 @@
-wikitools -- Package for working with MediaWiki wikis
+wikitools3 -- Package for working with MediaWiki wikis
 
 Requirements
 ------------
@@ -13,7 +13,7 @@ Requirements
 Installation
 ------------
 
-  * Run "python setup.py install" or copy the wikitools directory
+  * Run "python setup.py install" or copy the wikitools3 directory
     to an appropriate Python module directory.
   * An exe installer for Windows is also available (should be run as an 
     administrator to avoid errors)
@@ -65,8 +65,8 @@ To make a simple query:
 ```python
 #!/usr/bin/python
 
-from wikitools import wiki
-from wikitools import api
+from wikitools3 import wiki
+from wikitools3 import api
 
 # create a Wiki object
 site = wiki.Wiki("http://my.wikisite.org/w/api.php") 
@@ -107,8 +107,8 @@ pages in namespace 0 in a category:
 ```python
 #!/usr/bin/python
 
-from wikitools import wiki
-from wikitools import category
+from wikitools3 import wiki
+from wikitools3 import category
 
 site = wiki.Wiki("http://my.wikisite.org/w/api.php") 
 site.login("username", "password")
