+++ wikitools3/__init__.py
@@ -1,24 +1,25 @@
-﻿# -*- coding: utf-8 -*-
+﻿#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 # -*- coding: utf-8 -*-
 # Copyright 2008-2013 Alex Zaddach (mrzmanwiki@gmail.com)
 
-# This file is part of wikitools.
-# wikitools is free software: you can redistribute it and/or modify
+# This file is part of wikitools3.
+# wikitools3 is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
- 
-# wikitools is distributed in the hope that it will be useful,
+
+# wikitools3 is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
- 
+
 # You should have received a copy of the GNU General Public License
-# along with wikitools.  If not, see <http://www.gnu.org/licenses/>.
+# along with wikitools3.  If not, see <http://www.gnu.org/licenses/>.
 __all__ = ["wiki", "api", "page", "category", "user", "pagelist", "wikifile"]
-from wiki import *
-from api import *
-from page import *
-from category import *
-from user import *
-from wikifile import *
+from wikitools3.api import *
+from wikitools3.category import *
+from wikitools3.page import *
+from wikitools3.user import *
+from wikitools3.wiki import *
+from wikitools3.wikifile import *
